Added examples for PSTPairs, path graphs; edited GitHub actions
Began tests to see whether preliminary algo. results agree with theoretical results on hypercubes and path graphs. Plan to refactor the PSTPairs struct to show not only all PST qubit pairs but also the optimal transfer achieved over every pair. On the topic of path graphs, changing the optimization parameters in the examples affects maximum fidelity results found pairs in some of the higher-order path graphs. The IPNewton solver is getting stuck at local maxima and not checking the entire desired optimization range...how to fix this? (Additionally, made a job to test Julia v1.8 in addition to 1.11 and pre., since compat in Project.toml is 1.8+.)