Added examples for pst_pairs and path graphs
Began testing to see whether results from the preliminary algo. agree with theoretical results on hypercubes and path graphs. Plan to refactor the PSTPairs struct in the future to show not only all PST qubit pairs but also the optimal transfer achieved over every pair. On the topic of path graphs, changing the optimization parameters in the examples affects maximum fidelity results found pairs in some of the higher-order path graphs. It seems the IPNewton solver might be getting stuck at local maxima and not checking the entire desired optimization range...how to fix this?